Refactor: Create test module and move SemanticsExt
- Create a new `:carbon:test` module to hold testing utilities.
- Move SemanticsExt to this new module.
- Update usages of SemanticsExt to the new location.
- Add a README to the `:carbon:test` module.
