+++ carbon/src/commonTest/kotlin/com/gabrieldrn/carbon/radiobutton/RadioButtonTest.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.runComposeUiTest
 import com.gabrieldrn.carbon.common.selectable.SelectableInteractiveState
-import com.gabrieldrn.carbon.semantics.assertIsReadOnly
+import com.gabrieldrn.carbon.test.semantics.assertIsReadOnly
 import kotlin.test.Test
 import kotlin.test.assertTrue
