+++ carbon/src/commonTest/kotlin/com/gabrieldrn/carbon/toggle/ToggleTest.kt
@@ -37,7 +37,7 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.runComposeUiTest
 import androidx.compose.ui.unit.dp
-import com.gabrieldrn.carbon.semantics.isReadOnly
+import com.gabrieldrn.carbon.test.semantics.isReadOnly
 import kotlin.test.Test
 import kotlin.test.assertTrue
